# perf: Implement count pushdown for parquet