--- a/src/daft-parquet/src/read.rs
+++ b/src/daft-parquet/src/read.rs
@@ -1021,6 +1021,7 @@ pub async fn read_parquet_metadata_bulk(
     all_metadatas.into_iter().collect::<DaftResult<Vec<_>>>()
 }
 
+/// Optimized for count pushdowns: we can get the count from metadata without reading all data.
 pub async fn stream_parquet_count_pushdown(
     url: &str,
     io_client: Arc<IOClient>,
@@ -1028,7 +1029,6 @@ pub async fn stream_parquet_count_pushdown(
     field_id_mapping: Option<Arc<BTreeMap<i32, Field>>>,
     aggregation: &ExprRef,
 ) -> DaftResult<BoxStream<'static, DaftResult<RecordBatch>>> {
-    // For count pushdown, we can get the count from metadata without reading all data.
     let parquet_metadata =
         read_parquet_metadata(url, io_client, io_stats, field_id_mapping.clone()).await?;