## Changes Made

https://github.com/Eventual-Inc/Daft/pull/4969 implemented count pushdowns for lance. We can do the same for parquet and read only parquet metadata. For example, I have a bucket in `s3://desmond-test/big.parquet/` with the following setup:
- 161 parquet files
- Each parquet file has `256000` rows of strings with 1024 random characters, giving a total size of 1.2GB per file
- The footer len itself is only `2247` bytes

Before, doing a count takes ~28s:
```
In [2]: %time daft.read_parquet("s3://desmond-test/big.parquet").count().show()
╭──────────╮
│ count    │
│ ---      │
│ UInt64   │
╞══════════╡
│ 41216000 │
╰──────────╯

(Showing first 1 of 1 rows)
CPU times: user 1min 4s, sys: 1min 46s, total: 2min 50s
Wall time: 28.2 s
```

After it takes ~4s:
```
In [2]: %time daft.read_parquet("s3://desmond-test/big.parquet").count().show()
╭──────────╮
│ count    │
│ ---      │
│ UInt64   │
╞══════════╡
│ 41216000 │
╰──────────╯

(Showing first 1 of 1 rows)
CPU times: user 301 ms, sys: 94.2 ms, total: 395 ms
Wall time: 3.89 s
```